+++ src/components/EditionCard/EditionCard.tsx
@@ -71,6 +71,7 @@ export const EditionCard: React.FC<{
         layout="row"
         imageProps={{
           src: coverUrl,
+          fallbackSrc: PLACEHOLDER_COVER_LINK,
           alt:
             coverUrl === PLACEHOLDER_COVER_LINK
               ? "Placeholder Cover"