+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Prerelease]
 
 - Replace Cucumber with Playwright Page Object Model tests
+- Update GH action to run tests against local.nypl.org and QA for prod PRs
 
 ## [0.18.10]