+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Remove survey banner
 - Replace Cucumber with Playwright Page Object Model tests
 - Update GH action to run tests against local.nypl.org and QA for prod PRs
+- Add GTM scripts
 
 ## [0.18.10]